gpu_bridge: use size_t for readback clamp (fix x86 narrowing under /WX)

+++ src/windows-emulator/devices/gpu_bridge.cpp
@@ -470,7 +470,7 @@ namespace sogen
 
                 // output_buffer_length is guest-controlled; the response is a small fixed-size struct, so
                 // bound the staging allocation instead of trusting the declared length.
-                std::vector<std::byte> properties(std::min<uint64_t>(context.output_buffer_length, max_readback_bytes));
+                std::vector<std::byte> properties(std::min<size_t>(context.output_buffer_length, max_readback_bytes));
                 const int32_t result = this->vulkan_.get_physical_device_properties(request.physical_device, properties.data(),
                                                                                     properties.size(), win_emu.process.is_wow64_process);
                 if (result != 0)
@@ -1185,7 +1185,7 @@ namespace sogen
 
                 // output_buffer_length is guest-controlled; the response is a small fixed-size struct, so
                 // bound the staging allocation instead of trusting the declared length.
-                std::vector<std::byte> properties(std::min<uint64_t>(context.output_buffer_length, max_readback_bytes));
+                std::vector<std::byte> properties(std::min<size_t>(context.output_buffer_length, max_readback_bytes));
                 const int32_t result =
                     this->vulkan_.get_physical_device_memory_properties(request.physical_device, properties.data(), properties.size());
                 if (result != 0)
@@ -3050,7 +3050,7 @@ namespace sogen
 
                 // output_buffer_length is guest-controlled; the response is a small fixed-size struct, so
                 // bound the staging allocation instead of trusting the declared length.
-                std::vector<std::byte> caps(std::min<uint64_t>(context.output_buffer_length, max_readback_bytes));
+                std::vector<std::byte> caps(std::min<size_t>(context.output_buffer_length, max_readback_bytes));
                 const int32_t result = this->vulkan_.get_surface_capabilities(request.physical_device, request.surface, window_width,
                                                                               window_height, caps.data(), caps.size());
                 if (result != 0)
